+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # CHANGELOG
 
+
+**v0.2.3 - 05-14-2026**
+---
+
+- Added Czech translation
+- Added drag-and-drop to move messages to another folder
+- Added cross account copy/move mail - [#108](https://github.com/hkdb/aerion/issues/108)
+- Added draggable recipients in composer - [#111](https://github.com/hkdb/aerion/issues/111)
+- Added auto-commit recipient on lost focus - [#85](https://github.com/hkdb/aerion/issues/85)
+- Added composer del/backspace guard to prevent accidental message delete
+- Fixed detached composer system theme detection - [#153](https://github.com/hkdb/aerion/issues/153)
+- Fixed launch flow - [#154](https://github.com/hkdb/aerion/issues/154)
+- Fixed dark theme rendering - [#155](https://github.com/hkdb/aerion/issues/155)
+- Added unread count update after background sync to ensure accuracy
+
+
 **v0.2.2 - 05-09-2026**
 ---